Enhance event handling and menu management
- Added new command options to settings.json for improved functionality.
- Updated Event.java to register event listeners upon instantiation and suppress warnings for potential null values.
- Refactored Core.java to improve readability and maintainability, including changes to scoreboard management and race runner instantiation.
- Modified Main.java to standardize file path constants and improve JSON handling.
- Improved CustomMenu.java to encapsulate inventory management and item handling.
- Refined Menus.java for better race list and ranking display logic, ensuring proper inventory setup.
- Enhanced Race_Core.java with better handling of race states and player interactions.
- Cleaned up Race_Package.java by removing unused imports and variables.
- Updated Race_Runner.java to streamline vehicle handling and registration logic.
- Improved CreateJson.java for consistent file handling and logging.
- Fixed Speed_Meter.java to simplify distance calculation logic.

+++ src/main/java/waterpunch/atamamozi_d/plugin/tool/Scoreboaed/Speed_Meter.java
@@ -5,8 +5,8 @@ import org.bukkit.Location;
 public class Speed_Meter {
 
      public static double Meter(Location loc1, Location loc2) {
-          double X = Math.pow((double) loc1.getX() - (double) loc2.getX(), 2);
-          double Z = Math.pow((double) loc1.getZ() - (double) loc2.getZ(), 2);
+          double X = Math.pow(loc1.getX() - loc2.getX(), 2);
+          double Z = Math.pow(loc1.getZ() - loc2.getZ(), 2);
           return Math.sqrt(X + Z) * 1000;
      }
 }
